some little changes (rename, bug fix)
