+++ today/src/index.js
@@ -15,23 +15,10 @@ const theme = createTheme({
   
 });
 
-const message= [
-  "Сегодня у тебя всё получится!",
-  "Удача на твоей стороне!",
-  "Звёзды благоволят любым твоим начинаниям - начинай!",
-  "Сегодня произойдёт что-то невообразимое.",
-  "Это самый подходящий день, чтобы... отдохнуть.",
-  "Сегодня солнце будет светить исключительно для тебя.",
-  "Этот день обязательно принесёт что-то хорошее.",
-  "Когда, если не сегодня?"
-]
-let i = parseInt(Math.random()*message.length)
-//const PS = "* Все люди - как люди, а я на верблюде. И я вот не знаю, куда корректнее засунуть рандомайзер выбора сообщения. Положила в index, это было проще реализовать, но, возможно, я и не права... хм..."
-
 ReactDOM.render(
   <React.StrictMode>
     <ThemeProvider theme={theme}>
-      <App today={message[i]} />
+      <App />
     </ThemeProvider>
   </React.StrictMode>,
   document.getElementById('message')
